# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Cinema API"
  description: "This is a Cinema API where you can execute CRUD operations on 
  movies, actors, genres, sessions, customers and auditoriums"
servers:
  - url: http://127.0.0.1:5000
...
paths:
  /actors:
    get:
      tags:
        - "Actors"
      summary: "Get actors information"
      parameters:
        - name: "firstname"
          in: "query"
          description: "Firstname of actor"
          required: false
          schema:
            type: "string"
        - name: "lastname"
          in: "query"
          description: "Lastname of actor"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorsOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Actors"
      summary: "Create new actor"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'firstname' and 'lastname'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
...
  /genres:
    get:
      tags:
        - "Genres"
      summary: "Get genres information"
      parameters:
        - name: "genre"
          in: "query"
          description: "Genre name"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenresOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Genres"
      summary: "Create new genre"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'genre'"
  /genres/{id}:
    get:
      tags:
        - "Genres"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of genre"
          required: true
          schema:
            type: "integer"
      summary: "Get genre by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenreOut'
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Genre not found"
          content:
            application/json:
              example:
                message: "Genre not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Genres"
      summary: "Update genre's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of genre"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong genre's id"
          content:
            application/json:
              example:
                message: "Genre not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Genres"
      summary: "Delete genre"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the genre"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong genre's id"
          content:
            application/json:
              example:
                message: "Genre not found"
...
  /movies:
    get:
      tags:
        - "Movies"
      summary: "Get movies information"
      parameters:
        - name: "name"
          in: "query"
          description: "Movie name"
          required: false
          schema:
            type: "string"
        - name: "director_id"
          in: "query"
          description: "Director ID"
          required: false
          schema:
            type: "string"
        - name: "offset"
          in: "query"
          description: "Offset"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "Limit"
          required: false
          schema:
            type: "string"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviesOut'
    post:
      security:
        - bearerAuth: []
      tags:
        - "Movies"
      summary: "Create new movie"
      description: "This can only be done by the logged in admin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify name, description, year, month, day, director_id, genres and actors."
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
  /movies/{id}:
    get:
      tags:
        - "Movies"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of movie"
          required: true
          schema:
            type: "integer"
      summary: "Get movie by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieOut'
        '404':
          description: "Movie not found"
          content:
            application/json:
              example:
                message: "Movie not found"
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Movies"
      summary: "Update movie's information"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of movie session"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong movie's id"
          content:
            application/json:
              example:
                message: "Movie not found"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "Movies"
      summary: "Delete movie"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the movie"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong movies id"
          content:
            application/json:
              example:
                message: "Movie not found"
  /movies/changes/{id}:
    patch:
      security:
        - bearerAuth: []
      tags:
        - "Movies"
      summary: "Delete movie's genre or actor"
      description: "This can only be done by the logged in admin"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of movie session"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieSpecialPatchIn'
      responses:
        '200':
          description: "Successful Operation"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '403':
          description: "Admin access rights required"
          content:
            application/json:
              example:
                message: "Forbidden"
        '404':
          description: "Wrong movie's id"
          content:
            application/json:
              example:
                message: "Movie not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ActorOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        firstname:
          type: "string"
          example: "Emma"
        lastname:
          type: "string"
          example: "Stone"
        movies:
          type: "array"
          example: []
    ActorsOut:
      type: "array"
      items:
        $ref: '#/components/schemas/ActorOut'
    ActorIn:
      type: "object"
      properties:
        firstname:
          type: "string"
          example: "Emma"
        lastname:
          type: "string"
          example: "Stone"
...
    GenreOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        genre:
          type: "string"
          example: "comedy"
        movies:
          type: "array"
          example: []
    GenresOut:
      type: "array"
      items:
        $ref: '#/components/schemas/GenreOut'
    GenreIn:
      type: "object"
      properties:
        genre:
          type: "string"
          example: "drama"
...
    MovieOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 2
        release_date:
          type: "string"
          example: "Sun, 15 May 2022 14:30:00 GMT"
        name:
          type: "string"
          example: "Evil Dead"
        description:
          type: "string"
          example: "Cool movie"
        sessions:
          type: "array"
          example: []
        genres:
          type: "array"
          example: []
        actors:
          type: "array"
          example: []
    MoviesOut:
      type: "array"
      items:
        $ref: '#/components/schemas/MovieOut'
    MovieIn:
      type: "object"
      properties:
        name:
          type: "string"
          example: "Evil Dead"
        description:
          type: "string"
          example: "Cool movie"
        director_id:
          type: "integer"
          example: 2
        year:
          type: "integer"
          example: 2022
        month:
          type: "integer"
          example: 5
        day:
          type: "integer"
          example: 15
        genres:
          type: "array"
          example: ["horror", "slasher"]
        actors:
          type: "array"
          example: ["Bruce Campbell", "Someone Else"]
    MovieSpecialPatchIn:
      type: "object"
      properties:
        genres:
          type: "array"
          example: ["horror", "slasher"]
        actors:
          type: "array"
          example: ["Bruce Campbell", "Someone Else"]
...